pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "declarativedocs"
version = "0.1.0"
description = "Declarative Document Format (DDF) — YAML to documents compiler"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "WeAreBrain", email = "hello@wearebrain.com" },
]
keywords = ["ddf", "declarative", "documents", "pptx", "docx", "pdf", "yaml", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "python-pptx>=1.0",
    "PyYAML>=6.0",
    "lxml>=4.9",
]

[project.scripts]
ddf = "ddf.__main__:main"

[project.urls]
Homepage = "https://ddf.dev"
Repository = "https://github.com/declarativedocs/compiler-py"
Specification = "https://github.com/declarativedocs/spec"
